converter/model/requirements.txt
onnxruntime
Pillow